requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
gunicorn==22.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.2.0
requests==2.32.3
setuptools==68.0.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
